membuat InventoryApp dan mendaftakannya di android manifest

+++ app/src/main/java/com/example/ucproomdb/data/entity/Barang.kt
@@ -10,7 +10,7 @@ data class Barang(
     val idBarang :String,
     val nama :String,
     val deskripsi :String,
-    val harga :Int,
+    val harga :Double,
     val stok :Int,
     val namaSupplier :String,
 )
